requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
Werkzeug==3.0.1
WTForms==3.1.1
